Consolidate span_map with tracer_console
